agents/build-from-scratch/pyproject.toml
[project]
name = "build-from-scratch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ollama>=0.5.1",
    "smolagents[litellm,toolkit]>=1.19.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "python-dotenv>=1.1.1",
]
